Remove absolute host redirects.
Default nginx setting is to add $scheme$host(:$port) to all absolute
redirects. Similarly, for proxy_pass directives, the default behavior is
to rewrite Location: header (used for redirects) as if it was emitted
relative to the current server handling the request.

Main difference is that this unbreaks standalone set up, where redirects
used to be rewritten to https://localhost/$foo.
Such redirects then couldn't be rewritten by externally available
cms.lmio.lt server to point to correct host in the location header.